+++ cmd/goeland-server/goeland-front/src/router/index.ts
@@ -6,7 +6,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
-    { path: '/', redirect: '/documents' },
+    { path: '/', redirect: '/cases' },
     {
       path: '/documents',
       name: 'documents',